backend/requirements.txt
# Основные зависимости Django
Django>=4.2,<5.0
python-dotenv>=1.0.0

# WSGI сервер для production
gunicorn>=21.2.0

# Dev зависимости (опционально)
django-browser-reload>=1.8.0
